Add tcp server start function
tcp_connect: contains two functions tcp_server_init and tcp_server_start
tcp_server_start: creates a TCP socket with the specified port number
and waits for a connection request from the client.
After accepting a request from the client, connection between server
and client is established and the server waits for some data to be received from the client.

Signed-off-by: Ayush Kaura <ayushkaura@gmail.com>